requirements.txt
# deep learning
torchvision
torchmetrics
lightning
nflows

# data science / plotting
numpy
matplotlib
scipy
seaborn
scikit-learn
pandas
numba

# wandb/hydra
omegaconf
hydra-core
wandb
plotly # required for wandb (i think)

# tests
black
coverage
darglint
flake8
isort
pylint
pytest-cov
pytest

# misc
h5py
jupyter
tqdm
jupyterlab
python-gitlab
pyrootutils
pyjet

# workflow
snakemake

